+++ src/service/translation/translator.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class Translator(ABC):
+    @abstractmethod
+    def translate(self, content: list) -> list:
+        pass
