+++ work_space_tbd.py
@@ -42,228 +42,6 @@ os.environ['PYTHONIOENCODING'] = 'utf-8'
 os.environ["KMP_DUPLICATE_LIB_OK"] = "TRUE"  # 强制GPU版本cuda
 
 
-def download_youtube_video(video_id, fileNameAndPath):
-    from pytube import YouTube
-    YouTube(f'https://youtu.be/{video_id}', proxies=proxies).streams.first().download(filename=fileNameAndPath)
-
-
-def transcribeAudioEn(logger, path, modelName="base.en", language="en", srtFilePathAndName="VIDEO_FILENAME.srt"):
-    # 非静音检测阈值，单位为分贝，越小越严格
-    NOT_SILENCE_THRESHOLD_DB = -30
-
-    END_INTERPUNCTION = ["…", ".", "!", "?", ";"]
-    NUMBER_CHARACTERS = "0123456789"
-    # 确保简体中文
-    initial_prompt = None
-    if language == "zh":
-        initial_prompt = "简体"
-
-    model = WhisperModel(modelName, device="cuda", compute_type="float16", download_root="faster-whisper_models",
-                         local_files_only=False)
-    logger.info("Whisper model loaded.")
-
-    # faster-whisper
-    segments, _ = model.transcribe(audio=path, language=language, word_timestamps=True, initial_prompt=initial_prompt)
-    # 转换为srt的Subtitle对象
-    index = 1
-    subs = []
-    subtitle = None
-    for segment in segments:
-        for word in segment.words:
-            if subtitle is None:
-                subtitle = srt.Subtitle(index, datetime.timedelta(seconds=word.start),
-                                        datetime.timedelta(seconds=word.end), "")
-            finalWord = word.word.strip()
-            subtitle.end = datetime.timedelta(seconds=word.end)
-
-            # 一句结束。但是要特别排除小数点被误认为是一句结尾的情况。
-            if (finalWord[-1] in END_INTERPUNCTION) and not (len(finalWord) > 1 and finalWord[-2] in NUMBER_CHARACTERS):
-                pushWord = " " + finalWord
-                subtitle.content += pushWord
-                subs.append(subtitle)
-                index += 1
-                subtitle = None
-
-            else:
-                if subtitle.content == "":
-                    subtitle.content = finalWord
-                # 如果上一个字符是"."，则要考虑小数的可能性
-                elif finalWord[0] == ".":
-                    subtitle.content = subtitle.content + finalWord
-                elif len(subtitle.content) > 0 and subtitle.content[-1] == "." and finalWord[0] in NUMBER_CHARACTERS:
-                    subtitle.content = subtitle.content + finalWord
-                else:
-                    subtitle.content = subtitle.content + " " + finalWord
-
-    # 补充最后一个字幕
-    if subtitle is not None:
-        subs.append(subtitle)
-        index += 1
-
-    logger.info("Transcription complete.")
-
-    # 重新校准字幕开头，以字幕开始时间后声音大于阈值的第一帧为准
-    audio = wave.open(path, 'rb')
-    frameRate = audio.getframerate()
-    notSilenceThreshold = math.pow(10, NOT_SILENCE_THRESHOLD_DB / 20)
-    for sub in subs:
-        startTime = sub.start.total_seconds()
-        startFrame = int(startTime * frameRate)
-        endTime = sub.end.total_seconds()
-        endFrame = int(endTime * frameRate)
-
-        newStartTime = startTime
-        audio.setpos(startFrame)
-        readFrames = endFrame - startFrame
-        for i in range(readFrames):
-            frame = audio.readframes(1)
-            if not frame:
-                break
-            samples = struct.iter_unpack("<h", frame)
-            sampleVolumes = []  # 用于存储每个样本的音量值
-            for sample_tuple in samples:
-                # sample是一个样本值
-                # 调用calculate_volume函数计算样本的音量值，并将结果添加到sampleVolumes列表中
-                sample = sample_tuple[0]
-                sample_volume = abs(sample) / 32768
-                sampleVolumes.append(sample_volume)  # 将音量值添加到列表中
-            # 找出所有样本的音量值中的最大值
-            maxVolume = max(sampleVolumes)
-
-            if maxVolume > notSilenceThreshold:
-                newStartTime = startTime + i / frameRate
-                break
-
-        sub.start = datetime.timedelta(seconds=newStartTime)
-
-    content = srt.compose(subs)
-    with open(srtFilePathAndName, "w", encoding="utf-8") as file:
-        print(f"write to file {srtFilePathAndName}, file {file}")
-        file.write(content)
-
-    logger.info("SRT file created.")
-    logger.info("Output file: " + srtFilePathAndName)
-    return True
-
-
-def transcribeAudioZh(logger, path, modelName="base.en", language="en", srtFilePathAndName="VIDEO_FILENAME.srt"):
-    END_INTERPUNCTION = ["。", "！", "？", "…", "；", "，", "、", ",", ".", "!", "?", ";"]
-    ENGLISH_AND_NUMBER_CHARACTERS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-
-    model = WhisperModel(modelName, device="cuda", compute_type="float16", download_root="faster-whisper_models",
-                         local_files_only=False)
-    segments, _ = model.transcribe(audio=path, language="zh", word_timestamps=True, initial_prompt="简体")
-    index = 1
-    subs = []
-    for segment in segments:
-        subtitle = None
-        for word in segment.words:
-            if subtitle is None:
-                subtitle = srt.Subtitle(index, datetime.timedelta(seconds=word.start),
-                                        datetime.timedelta(seconds=word.end), "")
-            finalWord = word.word.strip()
-            subtitle.end = datetime.timedelta(seconds=word.end)
-
-            # 排除英文字母+. 情况
-            if ((finalWord[-1] in END_INTERPUNCTION and
-                 not (finalWord[-1] == "." and len(finalWord) > 1 and
-                      finalWord[-2] in ENGLISH_AND_NUMBER_CHARACTERS)) or
-                    (subtitle is not None and len(subtitle.content) > 20)):
-                if not ((finalWord[-1] == "." and len(finalWord) > 1 and
-                         finalWord[-2] in ENGLISH_AND_NUMBER_CHARACTERS) or
-                        (subtitle is not None and len(subtitle.content) > 20)):
-                    pushWord = finalWord[:-1]
-                else:
-                    pushWord = finalWord
-                subtitle.content += pushWord
-                subs.append(subtitle)
-                index += 1
-                subtitle = None
-            else:
-                subtitle.content += finalWord
-
-        if subtitle is not None:
-            subs.append(subtitle)
-            index += 1
-
-    content = srt.compose(subs)
-    with open(srtFilePathAndName, "w", encoding="utf-8") as file:
-        file.write(content)
-
-
-def srtSentanceMerge(logger, sourceSrtFilePathAndName, OutputSrtFilePathAndName):
-    srtContent = open(sourceSrtFilePathAndName, "r", encoding="utf-8").read()
-    subGenerator = srt.parse(srtContent)
-    subList = list(subGenerator)
-    if len(subList) == 0:
-        print("No subtitle found.")
-        return False
-
-    subPorcessingIndex = 1
-    subItemList = []
-    subItemProcessing = None
-    for subItem in subList:
-        dotIndex = subItem.content.rfind('.')
-        exclamationIndex = subItem.content.rfind('!')
-        questionIndex = subItem.content.rfind('?')
-        endSentenceIndex = max(dotIndex, exclamationIndex, questionIndex)
-
-        # 异常情况，句号居然在中间
-        if endSentenceIndex != -1 and endSentenceIndex != len(subItem.content) - 1:
-            logString = f"Warning: Sentence (index:{endSentenceIndex}) not end at the end of the subtitle.\n"
-            logString += f"Content: {subItem.content}"
-            logger.info(logString)
-
-        # 以后一个字幕，直接拼接送入就可以了
-        if subItem == subList[-1]:
-            if subItemProcessing is None:
-                subItemProcessing = copy.copy(subItem)
-                subItemList.append(subItemProcessing)
-                break
-            else:
-                subItemProcessing.end = subItem.end
-                subItemProcessing.content += subItem.content
-                subItemList.append(subItemProcessing)
-                break
-
-        # 新处理一串字符，则拷贝
-        if subItemProcessing is None:
-            subItemProcessing = copy.copy(subItem)
-            subItemProcessing.content = ''  # 清空内容是为了延续后面拼接的逻辑
-
-        subItemProcessing.index = subPorcessingIndex
-        subItemProcessing.end = subItem.end
-        subItemProcessing.content += subItem.content
-        # 如果一句话结束了，就把这一句话送入处理
-        if endSentenceIndex == len(subItem.content) - 1:
-            subItemList.append(subItemProcessing)
-            subItemProcessing = None
-            subPorcessingIndex += 1
-
-    srtContent = srt.compose(subItemList)
-    # 如果打开错误则返回false
-    with open(OutputSrtFilePathAndName, "w", encoding="utf-8") as file:
-        file.write(srtContent)
-
-
-def srt_to_text(srt_path):
-    with open(srt_path, "r", encoding="utf-8") as file:
-        lines = [line.rstrip() for line in file.readlines()]
-    text = ""
-    for line in lines:
-        line = line.replace('\r', '')
-        if line.isdigit():
-            continue
-        if line == "\n":
-            continue
-        if line == "":
-            continue
-        if re.search('\\d{2}:\\d{2}:\\d{2},\\d{3} --> \\d{2}:\\d{2}:\\d{2},\\d{3}', line):
-            continue
-        text += line + '\n'
-    return text
-
-
 def googleTrans(texts):
     if PROXY == "":
         client = Translate()
@@ -503,33 +281,6 @@ def srtToVoiceEdge(logger, srtFileNameAndPath, outputDir, character="zh-CN-Xiaoy
     return True
 
 
-def zhVideoPreview(logger, videoFileNameAndPath, voiceFileNameAndPath, insturmentFileNameAndPath, srtFileNameAndPath,
-                   outputFileNameAndPath):
-    # 从moviepy.editor导入VideoFileClip的创建音-视频剪辑
-    video_clip = VideoFileClip(videoFileNameAndPath)
-
-    # 加载音频
-    voice_clip = None
-    if (voiceFileNameAndPath is not None) and os.path.exists(voiceFileNameAndPath):
-        voice_clip = AudioFileClip(voiceFileNameAndPath)
-    insturment_clip = None
-    if (insturmentFileNameAndPath is not None) and os.path.exists(insturmentFileNameAndPath):
-        insturment_clip = AudioFileClip(insturmentFileNameAndPath)
-
-    # 组合音频剪辑
-    final_audio = None
-    if voiceFileNameAndPath is not None and insturmentFileNameAndPath is not None:
-        final_audio = CompositeAudioClip([voice_clip, insturment_clip])
-    elif voiceFileNameAndPath is not None:
-        final_audio = voice_clip
-    elif insturmentFileNameAndPath is not None:
-        final_audio = insturment_clip
-
-    video_clip = video_clip.set_audio(final_audio)
-    video_clip.write_videofile(outputFileNameAndPath, codec='libx264', audio_codec='aac',
-                               remove_temp=True, logger=None)
-    video_clip.close()
-    return True
 
 
 def voiceConnect(logger, sourceDir, outputAndPath, warningFilePath):
